--- src/components/setting/sync/Enum.ts
@@ -1,2 +1,0 @@
-// change this in android/app/build.gradle
-export const RedirectURL = 'com.noxplayer://oauth';
