+++ src/components/setting/sync/Enums.ts
@@ -1,0 +1,1 @@
+export const RedirectUrl = 'com.noxplayer://oauthredirect';
