+++ src/components/setting/sync/GiteeAuth.ts
@@ -1,23 +1,22 @@
 import { authorize } from 'react-native-app-auth';
-import { getArrayBufferForBlob } from 'react-native-blob-jsi-helper';
 
 // eslint-disable-next-line import/no-unresolved
-import { GITHUB_KEY, GITHUB_SECRET } from '@env';
+import { GITEE_KEY, GITEE_SECRET } from '@env';
 import { logger } from '@utils/Logger';
 import GenericSyncButton from './GenericSyncButton';
-import { checkAuthentication, noxBackup, noxRestore } from '@utils/sync/Github';
-import { RedirectURL } from '../Enum';
+import { checkAuthentication, noxBackup, noxRestore } from '@utils/sync/Gitee';
+import { RedirectUrl } from './Enums';
 
 const config = {
-  redirectUrl: RedirectURL,
-  clientId: GITHUB_KEY,
-  clientSecret: GITHUB_SECRET,
-  scopes: ['identity', 'repo', 'administration:write'],
+  redirectUrl: RedirectUrl,
+  redirectUri: RedirectUrl,
+  clientId: GITEE_KEY,
+  clientSecret: GITEE_SECRET,
+  scopes: ['projects', 'user_info'],
   additionalHeaders: { Accept: 'application/json' },
   serviceConfiguration: {
-    authorizationEndpoint: 'https://github.com/login/oauth/authorize',
-    tokenEndpoint: 'https://github.com/login/oauth/access_token',
-    revocationEndpoint: `https://github.com/settings/connections/applications/${GITHUB_KEY}`,
+    authorizationEndpoint: 'https://gitee.com/oauth/authorize',
+    tokenEndpoint: 'https://gitee.com/oauth/token',
   },
 };
 
@@ -29,7 +28,7 @@ export const getAuth = async (
 ) => {
   const authState = await authorize(config);
   if (authState.accessToken) {
-    logger.debug('github login successful');
+    logger.debug('gitee login successful');
     authToken = authState.accessToken;
     callback();
   } else {
@@ -37,30 +36,33 @@ export const getAuth = async (
   }
 };
 
-export const login = async (
+const login = async (
   callback: () => Promise<void> = async () => undefined,
   errorCallback = logger.error
 ) => {
   try {
-    if (!(await checkAuthentication(authToken))) {
-      logger.debug('Github token expired, need to log in');
+    if (!(await checkAuthentication())) {
+      logger.debug('gitee token expired, need to log in');
       await getAuth(callback, errorCallback);
     } else {
       callback();
     }
     return true;
   } catch (e) {
-    logger.debug('Github fail');
+    logger.debug('gitee fail');
     errorCallback(e);
     return false;
   }
 };
 
-export default ({ restoreFromUint8Array }: NoxSyncComponent.GenericProps) =>
+const GiteeSyncButton = ({
+  restoreFromUint8Array,
+}: NoxSyncComponent.GenericProps) =>
   GenericSyncButton({
     restoreFromUint8Array,
     noxBackup: v => noxBackup(v, authToken),
-    noxRestore: () =>
-      noxRestore(authToken, async v => getArrayBufferForBlob(v)),
+    noxRestore: () => noxRestore(authToken),
     login,
   });
+
+export default GiteeSyncButton;
