+++ src/components/setting/sync/GithubAuth.ts
@@ -1,4 +1,4 @@
-import { useAuthRequest } from 'expo-auth-session';
+import { authorize } from 'react-native-app-auth';
 import { getArrayBufferForBlob } from 'react-native-blob-jsi-helper';
 
 // eslint-disable-next-line import/no-unresolved
@@ -6,11 +6,11 @@ import { GITHUB_KEY, GITHUB_SECRET } from '@env';
 import { logger } from '@utils/Logger';
 import GenericSyncButton from './GenericSyncButton';
 import { checkAuthentication, noxBackup, noxRestore } from '@utils/sync/Github';
-import { RedirectURL } from '../Enum';
+import { RedirectUrl } from './Enums';
 
 const config = {
-  redirectUrl: RedirectURL,
-  redirectUri: RedirectURL,
+  redirectUrl: RedirectUrl,
+  redirectUri: RedirectUrl,
   clientId: GITHUB_KEY,
   clientSecret: GITHUB_SECRET,
   scopes: ['identity', 'repo', 'administration:write'],
@@ -24,37 +24,44 @@ const config = {
 
 let authToken = '';
 
-const useLogin = () => {
-  const [, , promptAsync] = useAuthRequest(config, config.serviceConfiguration);
-  const login = async (
-    callback: () => Promise<void> = async () => undefined,
-    errorCallback = logger.error
-  ) => {
-    try {
-      if (!(await checkAuthentication())) {
-        logger.debug('github token expired, need to log in');
-        const result = await promptAsync();
-        if (result.type === 'success') {
-          authToken = result.params.code;
-        } else {
-          errorCallback('[Github] auth failed');
-        }
-      } else {
-        callback();
-      }
-      return true;
-    } catch (e) {
-      errorCallback(e);
-      return false;
-    }
-  };
-  return { login };
+export const getAuth = async (
+  callback = () => checkAuthentication(authToken).then(console.log),
+  errorHandling = logger.error
+) => {
+  const authState = await authorize(config);
+  if (authState.accessToken) {
+    logger.debug('github login successful');
+    authToken = authState.accessToken;
+    callback();
+  } else {
+    errorHandling('no response url returned. auth aborted by user.');
+  }
 };
+
+export const login = async (
+  callback: () => Promise<void> = async () => undefined,
+  errorCallback = logger.error
+) => {
+  try {
+    if (!(await checkAuthentication(authToken))) {
+      logger.debug('Github token expired, need to log in');
+      await getAuth(callback, errorCallback);
+    } else {
+      callback();
+    }
+    return true;
+  } catch (e) {
+    logger.debug('Github fail');
+    errorCallback(e);
+    return false;
+  }
+};
+
 export default ({ restoreFromUint8Array }: NoxSyncComponent.GenericProps) =>
   GenericSyncButton({
     restoreFromUint8Array,
     noxBackup: v => noxBackup(v, authToken),
     noxRestore: () =>
       noxRestore(authToken, async v => getArrayBufferForBlob(v)),
-    useLogin,
+    login,
   });
